AESDefinitions: Added paddedRoundKeys_t type to hold round keys
The fields of a packed union must be of the same length, so the expandedKey_t union needs the round
key vector to be padded to be the same length as the expanded key blocks.

Given a variable of the expandedKey_t type called "expk", the individual round keys can be accessed as follows:

    roundKey_i = expk.roundKeys.keys[i];

The union is used so the expandKey module can more easily access the key schedule by
cipher-key-sized blocks, and by 4-byte columns.