Makefile, EncoderDecoderTestBench: Moved HVL testbench to new folder
Since the makefile was changed to compile by wildcard, having the EncoderDecoderTestBench.sv file in
the same directory as other testbenches caused compile errors when compiling in puresim mode. This
file was moved to a separate directory, and is now only included in compilation mode if not
compiling in puresim.